+++ srv/srvconfig/pagination.go
@@ -1,4 +1,4 @@
-package srvoption
+package srvconfig
 
 import (
 	"fmt"
